+++ CHANGELOG.md
@@ -5,6 +5,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+Upcoming 2.0 changes
+
+- [ ] Toggle base64 encoding (constructor level) in Asymmetric/Symmetric
+- [ ] Removed Deprecated Features
+- [ ] Make min 7.4
+- [ ] Adjust Hybrid to toggle base64 instead of decode
+
+## [1.2.0] - 2020-11-22
+
+### Added
+
+- Added Hybrid Encryption
+- Added PrivateKey::bits 
+- Added PublicKey::bits
+- Add check data length on encryption against key size limit
+
+### Changed
+
+- Changed boundaries are no longer added by default in Asymmetric Encryption
+
 ## [1.1.0] - 2020-11-21
 
 ### Changed